Revamp Flink IT by random port and merge tests
